requirements.txt
python-dotenv
google-genai
litellm
openai